Extract stack experience details dialog into standalone files
